requirements.txt
django==3.2
django-filter==2.4.0
django-environ==0.4.5
drf-yasg==1.20.0
djangorestframework==3.12.4
mysqlclient==2.0.3
